+++ src/drive/web/modules/filelist/FileListBody.jsx
@@ -1,12 +1,36 @@
 import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import Alerter from 'cozy-ui/react/Alerter'
-
-import AsyncFileListRows from 'drive/web/modules/filelist/AsyncFileListRows'
-import AddFolder from 'drive/web/modules/filelist/AddFolder'
+import Oops from 'components/Error/Oops'
+import { EmptyDrive, EmptyTrash } from 'components/Error/Empty'
+import AsyncBoundary from 'drive/web/modules/navigation/AsyncBoundary'
+import FileListRowsPlaceholder from './FileListRowsPlaceholder'
+import FileListRows from './FileListRows'
+import AddFolder from './AddFolder'
 
 import styles from 'drive/styles/filelist'
 
 const toggle = (flag, state, props) => ({ [flag]: !state[flag] })
+
+const EmptyContent = props => {
+  const { isTrashContext, canUpload } = props
+  if (isTrashContext && !props.params.folderId) {
+    return <EmptyTrash />
+  }
+  return <EmptyDrive canUpload={canUpload} />
+}
+
+EmptyContent.propTypes = {
+  isTrashContext: PropTypes.bool,
+  canUpload: PropTypes.bool,
+  params: PropTypes.object
+}
+
+EmptyContent.defaultProps = {
+  isTrashContext: false,
+  canUpload: false,
+  params: {}
+}
 
 export default class FileListBody extends Component {
   state = {
@@ -32,6 +56,7 @@ export default class FileListBody extends Component {
 
   render() {
     const { showAddFolder } = this.state
+    const { files } = this.props
     return (
       <div className={styles['fil-content-body']}>
         {showAddFolder && (
@@ -40,7 +65,15 @@ export default class FileListBody extends Component {
             onAbort={this.abortAddFolder}
           />
         )}
-        <AsyncFileListRows {...this.props} isAddingFolder={showAddFolder} />
+        <AsyncBoundary>
+          {({ isLoading, isInError }) => {
+            if (isLoading) return <FileListRowsPlaceholder />
+            else if (isInError) return <Oops />
+            else if (files.length === 0 && !showAddFolder)
+              return <EmptyContent {...this.props} />
+            else return <FileListRows withSelectionCheckbox {...this.props} />
+          }}
+        </AsyncBoundary>
       </div>
     )
   }
